Switched relic experience from int to double

+++ src/main/java/it/hurts/sskirillss/relics/api/events/leveling/ExperienceAddEvent.java
@@ -12,9 +12,9 @@ import javax.annotation.Nullable;
 public class ExperienceAddEvent extends RelicEvent implements ICancellableEvent {
     @Getter
     @Setter
-    private int amount;
+    private double amount;
 
-    public ExperienceAddEvent(@Nullable LivingEntity entity, ItemStack stack, int amount) {
+    public ExperienceAddEvent(@Nullable LivingEntity entity, ItemStack stack, double amount) {
         super(entity, stack);
 
         this.amount = amount;
